requirements.txt
torch
torchvision
opencv-python
pillow
matplotlib
mtcnn